LayoutTests/inspector/console/console-format-expected.txt

CONSOLE MESSAGE: line 20: test,test2,,,test4,,,,,
CONSOLE MESSAGE: line 21: %o
CONSOLE MESSAGE: line 22: %O
CONSOLE MESSAGE: line 23: Test for zero "%f" in formatter
CONSOLE MESSAGE: line 53: /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
CONSOLE MESSAGE: line 54: /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
CONSOLE MESSAGE: line 53: /foo\\bar\sbaz/i
CONSOLE MESSAGE: line 54: /foo\\bar\sbaz/i
CONSOLE MESSAGE: line 53: test
CONSOLE MESSAGE: line 54: test
CONSOLE MESSAGE: line 53: test named "test"
CONSOLE MESSAGE: line 54: test named "test"
CONSOLE MESSAGE: line 53: [object HTMLParagraphElement]
CONSOLE MESSAGE: line 54: [object HTMLParagraphElement]
CONSOLE MESSAGE: line 53: function () { return 1; }
CONSOLE MESSAGE: line 54: function () { return 1; }
CONSOLE MESSAGE: line 53: function () {
        return 2;
    }
CONSOLE MESSAGE: line 54: function () {
        return 2;
    }
CONSOLE MESSAGE: line 53: 0.12
CONSOLE MESSAGE: line 54: 0.12
CONSOLE MESSAGE: line 53: http://webkit.org/
CONSOLE MESSAGE: line 54: http://webkit.org/
CONSOLE MESSAGE: line 53: null
CONSOLE MESSAGE: line 54: 
CONSOLE MESSAGE: line 53: undefined
CONSOLE MESSAGE: line 54: 
CONSOLE MESSAGE: line 53: [object Attr]
CONSOLE MESSAGE: line 54: [object Attr]
CONSOLE MESSAGE: line 53: [object Attr]
CONSOLE MESSAGE: line 54: [object Attr]
CONSOLE MESSAGE: line 53: [object Attr]
CONSOLE MESSAGE: line 54: [object Attr]
CONSOLE MESSAGE: line 53: [object Object]
CONSOLE MESSAGE: line 54: [object Object]
CONSOLE MESSAGE: line 53: NaN
CONSOLE MESSAGE: line 54: NaN
CONSOLE MESSAGE: line 53: Infinity
CONSOLE MESSAGE: line 54: Infinity
CONSOLE MESSAGE: line 53: -Infinity
CONSOLE MESSAGE: line 54: -Infinity
CONSOLE MESSAGE: line 53: test,test2,,,test4,,,,,
CONSOLE MESSAGE: line 54: test,test2,,,test4,,,,,
CONSOLE MESSAGE: line 53: [object Object]
CONSOLE MESSAGE: line 54: [object Object]
CONSOLE MESSAGE: line 53: function () {}
CONSOLE MESSAGE: line 54: function () {}
CONSOLE MESSAGE: line 53: [object Object]
CONSOLE MESSAGE: line 54: [object Object]
Tests that console logging dumps proper messages.

Array[10] console-format.html:20
Array[10] console-format.html:21
Array[10] console-format.html:22
Test for zero "0" in formatter console-format.html:23
/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i console-format.html:53
[/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\…?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i] console-format.html:54
globals[0]
/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
/foo\\bar\sbaz/i console-format.html:53
[/foo\\bar\sbaz/i] console-format.html:54
globals[1]
/foo\\bar\sbaz/i
test console-format.html:53
["test"] console-format.html:54
globals[2]
"test"
test named "test" console-format.html:53
["test named "test""] console-format.html:54
globals[3]
"test named "test""
<p id="p">Tests that console logging dumps proper messages.</p> console-format.html:53
[p#p] console-format.html:54
globals[4]
<p id="p">Tests that console logging dumps proper messages.</p>
function () { return 1; } console-format.html:53
[function] console-format.html:54
globals[5]
function () { return 1; }
function () {
        return 2;
    } console-format.html:53
[function] console-format.html:54
globals[6]
function () {
        return 2;
    }
0.12 console-format.html:53
[0.12] console-format.html:54
globals[7]
0.12
http://webkit.org/ console-format.html:53
["http://webkit.org/"] console-format.html:54
globals[8]
"http://webkit.org/"
null console-format.html:53
[null] console-format.html:54
globals[9]
null
undefined console-format.html:53
[undefined] console-format.html:54
globals[10]
undefined
attr="" console-format.html:53
[attr] console-format.html:54
globals[11]
attr=""
attr="value""value" console-format.html:53
[attr] console-format.html:54
globals[12]
attr="value""value"
id="x""x" console-format.html:53
[id] console-format.html:54
globals[13]
id="x""x"
Object {} console-format.html:53
[Object] console-format.html:54
globals[14]
Object {}
NaN console-format.html:53
[NaN] console-format.html:54
globals[15]
NaN
Infinity console-format.html:53
[Infinity] console-format.html:54
globals[16]
Infinity
-Infinity console-format.html:53
[-Infinity] console-format.html:54
globals[17]
-Infinity
["test", "test2", 4: "test4", foo: Object] console-format.html:53
[Array[10]] console-format.html:54
globals[18]
["test", "test2", undefined × 2, "test4", undefined × 5]
Object {} console-format.html:53
[Object] console-format.html:54
globals[19]
Object {}
[function] console-format.html:53
[Array[1]] console-format.html:54
globals[20]
[function () {}]
Object {bar: "bar", foo: "foo"} console-format.html:53
[Object] console-format.html:54
globals[21]
Object {bar: "bar", foo: "foo"}
